新增切面日志雏形 新增全局异常步骤调用

+++ src/main/java/com/example/demo/Service/Impl/ApiTestCaseStepServiceImpl.java
@@ -115,8 +115,10 @@ public class ApiTestCaseStepServiceImpl implements ApiTestCaseStepService {
 
     @Override
     public void apiTestCaseStepsDelete(List<Integer> ids) {
-        for (int id : ids) {
-            apiTestCaseStepMapper.deleteStepByStepId(id);
+        if (!ids.isEmpty()) {
+            for (int id : ids) {
+                apiTestCaseStepMapper.deleteStepByStepId(id);
+            }
         }
     }
